# campus-recruitment-plan/校园招聘接口文档包/OpenAPI接口草案.yaml
openapi: 3.1.0
info:
  title: 校园智能招聘与面试预约平台 API
  version: 1.0.0
  description: 面向学生、企业 HR、管理员的校园招聘与面试预约平台 OpenAPI 草案。
servers:
- url: http://localhost:8080/api
  description: 本地开发环境
- url: https://api.example.com/api
  description: 生产环境占位
...
security:
- bearerAuth: []
paths:
  /auth/register:
    post:
      tags:
      - Auth
      summary: 注册学生或企业账号
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRegisterResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
  /auth/login:
    post:
      tags:
      - Auth
      summary: 用户登录
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLoginResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /auth/logout:
    post:
      tags:
      - Auth
      summary: 用户登出
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /auth/me:
    get:
      tags:
      - Auth
      summary: 获取当前登录用户
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCurrentUser'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /auth/permissions:
    get:
      tags:
      - Auth
      summary: 获取当前用户权限列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePermissionList'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
...
  /company/applications:
    get:
      tags:
      - Application
      summary: 企业投递列表
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseApplicationPage'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: jobId
        in: query
        required: false
        description: 岗位 ID
        schema:
          type: integer
          format: int32
      - name: status
        in: query
        required: false
        description: 投递状态
        schema:
          type: string
      - name: keyword
        in: query
        required: false
        description: 学生姓名/学校/专业
        schema:
          type: string
      - *id001
      - *id002
  /company/applications/{id}/status:
    put:
      tags:
      - Application
      summary: 企业修改投递状态
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseApplication'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录或 Token 失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 业务冲突，例如重复投递或重复预约
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 系统异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        description: 投递 ID
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationStatusUpdateRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserType:
      type: string
      enum:
      - STUDENT
      - COMPANY
      - ADMIN
...
    ApplicationStatus:
      type: string
      enum:
      - DELIVERED
      - VIEWED
      - INTERVIEW_INVITED
      - BOOKED
      - REJECTED
      - CANCELED
      - FINISHED
...
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 30002
        message:
          type: string
          example: 重复投递
        data:
          nullable: true
...
    RegisterRequest:
      type: object
      required:
      - username
      - password
      - confirmPassword
      - userType
      properties:
        username:
          type: string
          example: zhangsan
        password:
          type: string
          format: password
          example: '123456'
        confirmPassword:
          type: string
          format: password
          example: '123456'
        userType:
          $ref: '#/components/schemas/UserType'
        phone:
          type: string
          example: '13800000000'
        email:
          type: string
          format: email
          example: test@example.com
    RegisterResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          example: 10001
        username:
          type: string
          example: zhangsan
        userType:
          $ref: '#/components/schemas/UserType'
    LoginRequest:
      type: object
      required:
      - username
      - password
      - userType
      properties:
        username:
          type: string
          example: zhangsan
        password:
          type: string
          format: password
          example: '123456'
        userType:
          $ref: '#/components/schemas/UserType'
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: mock-jwt-token
        expireTime:
          type: integer
          format: int64
          example: 1770000000000
        userId:
          type: integer
          format: int64
          example: 10001
        username:
          type: string
          example: zhangsan
        nickname:
          type: string
          example: 张三
        userType:
          $ref: '#/components/schemas/UserType'
        roles:
          type: array
          items:
            type: string
          example:
          - STUDENT
        permissions:
          type: array
          items:
            type: string
          example:
          - resume:upload
          - application:create
    CurrentUser:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
        userType:
          $ref: '#/components/schemas/UserType'
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
    PermissionList:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
...
    ApplicationStatusUpdateRequest:
      type: object
      required:
      - status
      properties:
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        reason:
          type: string
    Application:
      type: object
      properties:
        id:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        studentName:
          type: string
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        jobId:
          type: integer
          format: int64
        jobTitle:
          type: string
        resumeId:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        applyTime:
          type: string
          format: date-time
...
    ApplicationPage:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        total:
          type: integer
          format: int64
          example: 100
        pageNum:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
...
    ApiResponseRegisterResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/RegisterResponse'
    ApiResponseLoginResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/LoginResponse'
    ApiResponseCurrentUser:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/CurrentUser'
    ApiResponsePermissionList:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/PermissionList'
...
    ApiResponseApplication:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/Application'
...
    ApiResponseApplicationPage:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/ApplicationPage'
...